% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpk.R
\name{cpk.plot}
\alias{cpk.plot}
\title{Plot Cpk}
\usage{
cpk.plot(dados, n, m, LSE, LIE, dist, B0 = 1000, CI = 0.95)
}
\arguments{
\item{data, n, m, LSE, LIE, dist, B0, CI}{Path to the input file}
}
\value{
tabela
}
\description{
This function plot the modified interval Cpk.
}
